poem-claude/requirements.txt
python-dotenv==1.0.1
anthropic==0.49.0
